Fixed index out of range error and exiting editor causing unresponsive terminal bug
Fixed index out of range error by ensuring that the index and indices slices always are within valid bounds. Fixed the unresponsive terminal bug where the terminal was unresponsive after opening a file with an editor and then closing it. That was fixed by configuring std correctly and rerunning lds after exiting the file.